requirements.txt
flask 
mysql-connector-python 
Pillow 
Werkzeug 
